Add warning re missing MPL import, fix check_html calls
