+++ nilearn/plotting/tests/test_js_plotting_utils.py
@@ -1,5 +1,8 @@
+import os
 import re
 import base64
+import webbrowser
+import tempfile
 
 import numpy as np
 import matplotlib
@@ -9,6 +12,11 @@ from nilearn import surface
 from nilearn.datasets import fetch_surf_fsaverage
 
 from numpy.testing import assert_warns
+try:
+    from lxml import etree
+    LXML_INSTALLED = True
+except ImportError:
+    LXML_INSTALLED = False
 
 
 def _normalize_ws(text):
@@ -201,6 +209,77 @@ def test_mesh_to_plotly():
     for i, key in enumerate(['_i', '_j', '_k']):
         assert np.allclose(
             js_plotting_utils.decode(plotly[key], '<i4'), triangles[:, i])
+
+
+def check_html(html, check_selects=True, plot_div_id='surface-plot'):
+    fd, tmpfile = tempfile.mkstemp()
+    try:
+        os.close(fd)
+        html.save_as_html(tmpfile)
+        with open(tmpfile) as f:
+            saved = f.read()
+        assert saved == html.get_standalone()
+    finally:
+        os.remove(tmpfile)
+    assert "INSERT" not in html.html
+    assert html.get_standalone() == html.html
+    assert html._repr_html_() == html.get_iframe()
+    assert str(html) == html.get_standalone()
+    assert '<meta charset="UTF-8" />' in str(html)
+    _check_open_in_browser(html)
+    resized = html.resize(3, 17)
+    assert resized is html
+    assert (html.width, html.height) == (3, 17)
+    assert 'width="3" height="17"' in html.get_iframe()
+    assert 'width="33" height="37"' in html.get_iframe(33, 37)
+    if not LXML_INSTALLED:
+        return
+    root = etree.HTML(html.html.encode('utf-8'),
+                      parser=etree.HTMLParser(huge_tree=True))
+    head = root.find('head')
+    assert len(head.findall('script')) == 5
+    body = root.find('body')
+    div = body.find('div')
+    assert ('id', plot_div_id) in div.items()
+    if not check_selects:
+        return
+    selects = body.findall('select')
+    assert len(selects) == 3
+    hemi = selects[0]
+    assert ('id', 'select-hemisphere') in hemi.items()
+    assert len(hemi.findall('option')) == 2
+    kind = selects[1]
+    assert ('id', 'select-kind') in kind.items()
+    assert len(kind.findall('option')) == 2
+    view = selects[2]
+    assert ('id', 'select-view') in view.items()
+    assert len(view.findall('option')) == 7
+
+
+def _open_mock(f):
+    print('opened {}'.format(f))
+
+
+def _check_open_in_browser(html):
+    wb_open = webbrowser.open
+    webbrowser.open = _open_mock
+    try:
+        html.open_in_browser(temp_file_lifetime=None)
+        temp_file = html._temp_file
+        assert html._temp_file is not None
+        assert os.path.isfile(temp_file)
+        html.remove_temp_file()
+        assert html._temp_file is None
+        assert not os.path.isfile(temp_file)
+        html.remove_temp_file()
+        html._temp_file = 'aaaaaaaaaaaaaaaaaaaaaa'
+        html.remove_temp_file()
+    finally:
+        webbrowser.open = wb_open
+        try:
+            os.remove(temp_file)
+        except Exception:
+            pass
 
 
 def test_to_color_strings():
