+++ nilearn/reporting/__init__.py
@@ -2,7 +2,8 @@
 Reporting code for nilearn
 """
 
-from .html_report import (ReportMixin, HTMLDocument,
+from .html_report import (ReportMixin, HTMLDocument, HTMLReport,
                           set_max_img_views_before_warning)
 
-__all__ = ['ReportMixin', 'HTMLDocument', 'set_max_img_views_before_warning']
+__all__ = ['ReportMixin', 'HTMLDocument', 'HTMLReport',
+           'set_max_img_views_before_warning']
