+++ nilearn/reporting/tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+Testing reporting code in nilearn
+"""
